actually scratch that

+++ src/main/scala/playground/Playground.scala
@@ -34,7 +34,7 @@ object Playground extends App {
     StructField("Displacement", DoubleType),
     StructField("Horsepower", LongType),
     StructField("Weight_in_lbs", LongType),
-    StructField("Acceleration", LongType),
+    StructField("Acceleration", DoubleType),
     StructField("Year", StringType),
     StructField("Origin", StringType)
   ))
